Add files via upload
This is an extra file off the main d3 file as extra functionality is still to be incorporated, but regardless is basically finished. Single line on click to area chart is working alongside a working dynamic legend and title along with axis labels. Code also cleaned up on all files.