Add files via upload
Proposed final submit implementation. Onclick function finished with the final design for the stack. Script not touched however removed button feature from stack.js as it was causing too many issues and is just easier to stick with what is working. Line chart height should probably be lowered a bit so that the user can see the whole thing without having to scroll. There is also an issue with the line chart legend when the user clicks on the line to view a stack and then goes back to the line chart.